Fixes #184 (#217)
