+++ models/userdevice.js
@@ -7,11 +7,11 @@ var mongoose = require('mongoose'),
 
 var UserDeviceSchema = new Schema({
     owner: {type: ObjectId, required: true},
-    androidRegistration: {type: String, unique: true},
-    iosDeviceToken: {type: String, unique: true},
+    androidRegistration: {type: String},
+    iosDeviceToken: {type: String},
     deviceType: {type: String},
     deviceModel: {type: String},
-    deviceId: {type: String, unique: true},
+    deviceId: {type: String},
     globalLocation: {type: [Number], index: '2d'},
     globalAltitude: {type: Number},
     globalAccuracy: {type: Number},
